refactor & one shot prover
