+++ Cargo.toml
@@ -16,6 +16,13 @@ ark-ec = "0.4"
 ark-ed-on-bn254 = "0.4"
 ark-ff = "0.4"
 ark-serialize = "0.4"
+async-compatibility-layer = { git = "https://github.com/EspressoSystems/async-compatibility-layer", tag = "1.4.1", features = [
+    "logging-utils",
+] }
+async-std = "1.12.0"
+async-trait = "0.1.77"
+clap = { version = "4.4", features = ["derive", "env"] }
+cld = "0.5"
 ethers = { version = "2.0", features = ["ws"] }
 
 hotshot = { git = "https://github.com/EspressoSystems/hotshot", tag = "0.5.7.1" }
@@ -40,10 +47,13 @@ jf-relation = { git = "https://github.com/EspressoSystems/jellyfish", features =
   "std",
 ] }
 jf-utils = { git = "https://github.com/EspressoSystems/jellyfish" }
+snafu = "0.7.4"
 surf-disco = { git = "https://github.com/EspressoSystems/surf-disco", tag = "v0.4.6" }
 tide-disco = { git = "https://github.com/EspressoSystems/tide-disco", tag = "v0.4.6" }
+tracing = "0.1"
 bytesize = "1.3"
 itertools = "0.12"
 rand_chacha = "0.3"
 rand_distr = "0.4"
 serde = { version = "1.0.195", features = ["derive"] }
+url = "2.3"
