+++ sequencer/Cargo.toml
@@ -23,15 +23,13 @@ ark-bls12-381 = "0.4"
 ark-ed-on-bn254 = { workspace = true }
 ark-serialize = { workspace = true, features = ["derive"] }
 ark-std = { workspace = true }
-async-compatibility-layer = { git = "https://github.com/EspressoSystems/async-compatibility-layer", tag = "1.4.1", features = [
-    "logging-utils",
-] }
-async-std = "1.12.0"
-async-trait = "0.1.77"
+async-compatibility-layer = { workspace = true }
+async-std = { workspace = true }
+async-trait = { workspace = true }
 bincode = "1.3.3"
 bytesize = { workspace = true }
-clap = { version = "4.4", features = ["derive", "env"] }
-cld = "0.5"
+clap = { workspace = true }
+cld = { workspace = true }
 commit = { git = "https://github.com/EspressoSystems/commit" }
 contract-bindings = { path = "../contract-bindings" }
 derivative = "2.2"
@@ -67,15 +65,15 @@ jf-utils = { workspace = true } # TODO temporary: used only for test_rng()
 rand = "0.8.5"
 sequencer-utils = { path = "../utils" }
 serde = { workspace = true }
-snafu = "0.7.4"
+snafu = { workspace = true }
 surf-disco = { workspace = true }
 tide-disco = { workspace = true }
 toml = "0.8"
-tracing = "0.1"
+tracing = { workspace = true }
 typenum = { version = "1.15.0", default-features = false, features = [
     "no_std",
 ] }
-url = "2.3"
+url = { workspace = true }
 
 # Dependencies for feature `testing`
 hotshot-testing = { workspace = true, optional = true }
